+++ src/main/java/duke/exceptions/AuguryException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class AuguryException extends Exception {
     public AuguryException(String message) {
