+++ src/main/java/duke/exceptions/FileIOException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class FileIOException extends AuguryException {
     public FileIOException(String message) {
