+++ src/main/java/duke/exceptions/InvalidActionException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class InvalidActionException extends AuguryException {
     public InvalidActionException(String message) {
