+++ src/main/java/duke/exceptions/InvalidTaskCreationException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class InvalidTaskCreationException extends AuguryException {
     public InvalidTaskCreationException(String message) {
