+++ src/main/java/duke/tasks/DeadlineTask.java
@@ -1,22 +1,24 @@
-package tasks;
+package duke.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class EventTask extends Task {
+public class DeadlineTask extends Task {
     protected LocalDateTime time;
     static DateTimeFormatter TIME_DISPLAY_FORMAT = DateTimeFormatter.ofPattern("d MMM y, E, kk:mm");
 
-    public EventTask(String description, LocalDateTime time) {
+    public DeadlineTask(String description, LocalDateTime time) {
         super(description);
         this.time = time;
     }
-    public EventTask(String description, LocalDateTime time, boolean isDone) {
+
+    public DeadlineTask(String description, LocalDateTime time, boolean isDone) {
         super(description, isDone);
         this.time = time;
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " + time.format(TIME_DISPLAY_FORMAT) + ")";
+        return "[D]" + super.toString() + " (by: " + time.format(TIME_DISPLAY_FORMAT) + ")";
     }
 }
