+++ src/main/java/duke/tasks/EventTask.java
@@ -1,23 +1,23 @@
-package tasks;
+package duke.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class DeadlineTask extends Task {
+public class EventTask extends Task {
     protected LocalDateTime time;
     static DateTimeFormatter TIME_DISPLAY_FORMAT = DateTimeFormatter.ofPattern("d MMM y, E, kk:mm");
 
-    public DeadlineTask(String description, LocalDateTime time) {
+    public EventTask(String description, LocalDateTime time) {
         super(description);
         this.time = time;
     }
-
-    public DeadlineTask(String description, LocalDateTime time, boolean isDone) {
+    public EventTask(String description, LocalDateTime time, boolean isDone) {
         super(description, isDone);
         this.time = time;
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + time.format(TIME_DISPLAY_FORMAT) + ")";
+        return "[E]" + super.toString() + " (at: " + time.format(TIME_DISPLAY_FORMAT) + ")";
     }
 }
