+++ src/main/java/duke/tasks/TodoTask.java
@@ -1,4 +1,4 @@
-package tasks;
+package duke.tasks;
 
 public class TodoTask extends Task {
     public TodoTask(String description) {
